backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyMySQL==1.1.0
python-decouple==3.8
Pillow>=9.0.0
django-filter==23.5
PyJWT==2.8.0
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0 